% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_m.R
\encoding{UTF-8}
\name{forecast_m}
\alias{forecast_m}
\title{Estimation du total d'un mois sur base des quelques jours observés.}
\usage{
forecast_m(data, date, x, unite = 1, decimal = 0, cumul = FALSE, mod = "NULL")
}
\arguments{
\item{data}{Un objet de type data.frame contenant au minimum la serie numerique et un vecteur de type date.}

\item{date}{La date à partir de laquelle on calcule la performance MTD.}

\item{x}{Un vecteur de type numeric contenant les observations du phénomène à etudier.}

\item{unite}{Un valeur de type numeric, definissant l'ordre d'echelle d'affisage du resultat de la fonction. Sa valeur par defaut est \code{1}.}

\item{decimal}{Un objet de type numeric permettant de préciser le nombre d'entier à afficher après la virgule. Sa valeur par defaut est \code{0}.}

\item{cumul}{Un objet de type logic qui permet d'indiquer si le vecteur \code{x} contient des valeurs cumulées. Sa valeur par defaut est \code{FALSE}.}

\item{mod}{Un argument qui permet de modifier le mode de calcul de ce forecast en appliquant un model. Sa valeur par defaut est \code{NULL}.
Pour cette première version, on ne va pas travailler sur le cas où mod n'est pas \code{NULL}.}
}
\value{
Un vecteur de longieur 1.
}
\description{
Cette fonction permet estimer le total d'un mois sur base de MTD et de nombre de jour restant au courant de ce mois.
}
\examples{
x <- rnorm(102, mean = 50, sd = 6.3)
date <- seq.Date(as.Date("2022-12-01"), by = "day", length.out = 102)
df <- data.frame(date, x)
forecast_m(df, date = "2023-01-25", x = "x")

}
