% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iperform-package.R
\docType{package}
\name{iperform-package}
\alias{iperform}
\alias{iperform-package}
\title{iperform: Performances d'une série temporelle}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Un outil pour calculer les performances d’une série temporelle en une date ou sur une période bien spécifique. Il est plus destiné à l’analyse des données dans les domaines financier, bancaire, télécommunication ou marketing opérationnel.
}
\author{
\strong{Maintainer}: Patrick Iunga \email{patrick.ilunga@unikin.ac.cd}

Other contributors:
\itemize{
  \item  Iunga Buabua Patrick [copyright holder]
}

}
\keyword{internal}
