% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc-operateur_mobile.R
\docType{data}
\name{operateur_mobile}
\alias{operateur_mobile}
\title{operateur_mobile}
\format{
A data frame with 990 rows and 6 variables:
\describe{
\item{ date }{  Date, renseigne le mois où l’activité est observée }
\item{ Parc }{  integer, renseigne le nombre des clients actifs sur une periode de 90 jours }
\item{ Usage }{  numeric, renseigne l'usage consommé par l'ensemble des clients au cours du mois (La voix en minute et l'internet en megabyte) }
\item{ Revenu }{  numeric, renseigne le revenu généré par l'ensemble des clients après usage d'un service }
\item{ Service }{  character, renseigne le service mobile concerné (Voix, SMS, Internet) }
\item{ Operateur }{  character, renseigne l'operateur de téléphonie mobile concerné }
}
}
\source{
Source
}
\usage{
operateur_mobile
}
\description{
KPI Operateur Mobile.
}
\keyword{datasets}
