% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc-service_mobile.R
\docType{data}
\name{service_mobile}
\alias{service_mobile}
\title{service_mobile}
\format{
A data frame with 504 rows and 5 variables:
\describe{
\item{ date }{  Date, renseigne le mois où l’activité est observée }
\item{ Parc }{  integer, renseigne le nombre des clients actifs sur une periode de 90 jours }
\item{ Usage }{  numeric, renseigne l'usage consommé par l'ensemble des clients au cours du mois (La voix en minute, le SMS en nombre et l'internet en megabyte) }
\item{ Revenu }{  integer, renseigne le revenu généré par l'ensemble des clients après usage d'un service }
\item{ Service }{  character, renseigne le service mobile concerné (Voix, SMS, Internet) }
}
}
\source{
Source
}
\usage{
service_mobile
}
\description{
KPI Service Mobile.
}
\keyword{datasets}
