% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc-voix_mobile.R
\docType{data}
\name{voix_mobile}
\alias{voix_mobile}
\title{voix_mobile}
\format{
A data frame with 1003 rows and 4 variables:
\describe{
\item{ date }{  Date }
\item{ Parc }{  de type integer, le nombre de clients ayant utiliser le service voix dans le 30 jour  }
\item{ Usage }{  de type numeric, le nombre de minutes consommées par l'ensemble des clients }
\item{ Revenu }{  de type numeric, le revenu genéré par l'ensemble des clients après usage (sous un quelque tarif du service) }
}
}
\source{
Source
}
\usage{
voix_mobile
}
\description{
KPI Service Voix Mobile.
}
\keyword{datasets}
