% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc-voix_tel_mob.R
\docType{data}
\name{voix_tel_mob}
\alias{voix_tel_mob}
\title{voix_tel_mob}
\format{
A data frame with 1003 rows and 4 variables:
\describe{
\item{date}{Date}
\item{parc}{integer}
\item{usage}{numeric}
\item{revenu}{numeric}
}
}
\source{
Source
}
\usage{
voix_tel_mob
}
\description{
KPI service voix - Téléphonie mobile.
}
\keyword{datasets}
