% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iperform-package.R
\docType{package}
\name{iperform-package}
\alias{iperform}
\alias{iperform-package}
\title{iperform: Performance}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calcule la performance d'une série temporelle sur une période bien spécifique.
}
\author{
\strong{Maintainer}: Patrick ILUNGA \email{patrick.ilunga@unikin.ac.cd}

Other contributors:
\itemize{
  \item  ILUNGA BUABUA [copyright holder]
}

}
\keyword{internal}
