% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iperform-package.R
\docType{package}
\name{iperform-package}
\alias{iperform}
\alias{iperform-package}
\title{iperform: Time Series Performance}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tool to calculate the performance of a time series in a specific date or period. It is more intended for data analysis in the fields of finance, banking, telecommunications or operational marketing.
}
\author{
\strong{Maintainer}: Patrick Ilunga \email{patrick.ilunga@unikin.ac.cd}

Other contributors:
\itemize{
  \item  Ilunga Buabua Patrick [copyright holder]
}

}
\keyword{internal}
